pagination & search compa+uni

+++ SpringBootJPA_ISCmanagement_project/src/main/java/com/ISC/project/controller/CompanyController.java
@@ -5,6 +5,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ISC.project.exception.ResourseNotFoundException;
 import com.ISC.project.model.Company;
+import com.ISC.project.model.University;
 import com.ISC.project.payload.ResultRespon;
 import com.ISC.project.service.CompanyService;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -40,7 +45,12 @@ public class CompanyController {
 	public ResultRespon getCompany(@RequestParam("id") long id) {
 		List<Company> compa = new ArrayList<>();
 		compa.add(companyService.findById(id).orElseThrow(() -> new ResourseNotFoundException("not found company with id: " + id)));
-		return new ResultRespon(0,"Success",compa);
+		if(compa.isEmpty()) {
+			throw new ResourseNotFoundException("Not found company by id " + id);
+		}else {
+			return new ResultRespon(0,"Success",compa);
+		}
+		
 	}
 	
 	//post company
@@ -58,7 +68,7 @@ public class CompanyController {
 	}
 	
 	//update company
-	@PutMapping("/editCompany")
+	@PutMapping(value="/editCompany")
 	public ResultRespon editCompany(@RequestBody Company company,@RequestParam("id") long id) {
 		List<Company> compa = new ArrayList<Company>();
 		Company oldcompany = companyService.findById(id).orElseThrow(() -> new ResourseNotFoundException("not found company with id: "+ id));
@@ -81,10 +91,43 @@ public class CompanyController {
 	}
 	
 	//delete company
-	@DeleteMapping("/deleteCompany")
+	@DeleteMapping(value="/deleteCompany")
 	public ResultRespon deleteCompany(@RequestParam("id") long id) {
 		companyService.findById(id).orElseThrow(() -> new ResourseNotFoundException("not found company with id: "+ id));
-		this.companyService.delete(id);
-		return new ResultRespon(0,"Delete company id = " + id + " success");
+		if(this.companyService.findById(id).isEmpty()) {
+			throw new ResourseNotFoundException("Nothing to delete");
+		}else {
+			this.companyService.delete(id);
+			return new ResultRespon(0,"Delete company id = " + id + " success");
+		}
 	}
+	
+	//search by name
+	@GetMapping(value="/searchCompany")
+	public ResultRespon searchCompany(@RequestParam("keyWord") String keyWord) {
+		if(this.companyService.searchCompany(keyWord).isEmpty()) {
+			throw new ResourseNotFoundException("Not found company by keyword " + keyWord);
+		}else {
+			return new ResultRespon(0,"Success",this.companyService.searchCompany(keyWord));
+		}
+	}
+	
+	//Pagination
+		@GetMapping("/company/pagination")
+		public ResultRespon paginationUniversity(@RequestParam(name = "page", required = false, defaultValue = "1") Integer page,
+				@RequestParam(name = "size",required = false,defaultValue = "1") Integer size,
+				@RequestParam(name = "sort",required = false,defaultValue = "1") String sort) {
+			Sort sortable = null;
+			if(sort.equals("ASC")) {
+				sortable = Sort.by("nameCom").ascending();
+			}
+			if(sort.equals("DESC")) {
+				sortable = Sort.by("nameCom").descending();
+			}
+			Pageable pageable = PageRequest.of(page, size,sortable);
+			Page<Company> compa = companyService.findCompa(pageable);
+			List<Page<Company>> companies = new ArrayList<Page<Company>>();
+			companies.add(compa);
+			return new ResultRespon(0,"Success",companies);
+		}
 }
